Nomenclature cleaning
