+++ study/src/org/labkey/study/writer/SampleTypeDatasetDefinition.java
@@ -3,9 +3,9 @@ package org.labkey.study.writer;
 import org.labkey.api.writer.VirtualFile;
 import org.labkey.study.model.StudyImpl;
 
-// DatasetWriter actually writes all datasets (crf, assay, and sample type). This is a do-nothing writer to gets the
-// sample type dataset checkbox to show up in the UI.
-public class SampleTypeDatasetWriter implements InternalStudyWriter
+// DatasetDefinitionWriter actually writes all dataset definitions (study, assay, and sample type). This is a do-nothing writer to get the
+// sample type dataset definition checkbox to show up in the UI.
+public class SampleTypeDatasetDefinition implements InternalStudyWriter
 {
     @Override
     public String getDataType()
