+++ study/src/org/labkey/study/writer/StudyXmlWriter.java
@@ -35,8 +35,9 @@ import java.util.Calendar;
 // This writer is largely responsible for study.xml.  It constructs the StudyDocument (xml bean used to read/write study.xml)
 //  that gets added to the StudyExportContext, writes the top-level study attributes, and writes out the bean when it's complete.
 //  However, each top-level writer is responsible for their respective elements in study.xml -- VisitMapWriter handles "visits"
-//  element, DatasetWriter is responsible for "datasets" element, etc.  As a result, StudyXmlWriter must be invoked after all
-//  writers are done modifying the StudyDocument.  Locking the StudyDocument after writing it out helps ensure this ordering.
+//  element, DatasetDefinitionWriter and DatasetDataWriter are responsible for "datasets" element, etc.  As a result, StudyXmlWriter
+//  must be invoked after all writers are done modifying the StudyDocument.  Locking the StudyDocument after writing it out helps
+//  ensure this ordering.
 class StudyXmlWriter implements InternalStudyWriter
 {
     private static final String PROPERTIES_DIRECTORY = "properties";
